Change scenario and model loading
