--- models/dungeon1bombs/best.zip.json
@@ -1,1 +1,0 @@
-{"evaluation": 10.60958904109589, "iterations": 5177344}
