--- triforce_lib/models.json
@@ -1,54 +1,0 @@
-{
-    "models" :
-    [
-        {
-            "name" : "dungeon1nobeams",
-            "path" : "dungeon1nobeams/model.zip",
-            "action_space" : "attack-only",
-            "priority" : 0,
-            "levels" : 1,
-            "requires_enemies" : false,
-            "equipment_required" : [],
-
-            "training_scenario" : "dungeon1nobeams",
-            "iterations" : 1500000
-        },
-        {
-            "name" : "dungeon1beams",
-            "path" : "dungeon1beams/model.zip",
-            "action_space" : "attack-only",
-            "priority" : 1,
-            "levels" : 1,
-            "requires_enemies" : false,
-            "equipment_required" : ["beams"],
-
-            "training_scenario" : "dungeon1beams",
-            "iterations" : 1500000
-        },
-        {
-            "name" : "dungeon1bombs",
-            "path" : "dungeon1bombs/model.zip",
-            "action_space" : "directional-item",
-            "priority" : 2,
-            "levels" : 1,
-            "requires_enemies" : false,
-            "equipment_required" : ["bombs"],
-
-            "training_scenario" : "dungeon1bombs",
-            "iterations" : 1500000
-        },
-        {
-            "name" : "dungeon1boss",
-            "path" : "models/dungeon1boss.zip",
-            "action_space" : "attack-only",
-            "priority" : 10,
-            "levels" : 1,
-            "rooms" : ["0x35", "0x36"],
-            "requires_enemies" : true,
-            "equipment_required" : [],
-
-            "training_scenario" : "dungeon1boss",
-            "iterations" : 500000
-        }
-    ]
-}
