+++ triforce_lib/models.py
@@ -3,6 +3,8 @@ import os
 from typing import List, Optional
 from dataclasses import dataclass
 
+from stable_baselines3 import PPO
+
 @dataclass
 class ZeldaModelInfo:
     name : str
@@ -19,17 +21,17 @@ class ZeldaModelInfo:
     iterations : int
 
 class ZeldaModel(ZeldaModelInfo):
+    _models = {}
+    _model_infos = []
+
     def __init__(self, model_info : ZeldaModelInfo, model):
         super().__init__(**model_info.__dict__)
         self.model = model
 
-def load_model_info():
-    model_json_path = os.path.join(os.path.dirname(__file__), "models.json")
-
-    models = []
-    with open(model_json_path) as f:
-        models_json = json.load(f)
-        for model_json in models_json['models']:
+    @classmethod
+    def initialize(cls, models_json):
+        model_info = []
+        for model_json in models_json:
             if isinstance(model_json['levels'], int):
                 model_json['levels'] = [model_json['levels']]
 
@@ -42,6 +44,37 @@ def load_model_info():
             elif isinstance(model_json['rooms'], list):
                 model_json['rooms'] = [int(x, 16) for x in model_json['rooms']]
 
-            models.append(ZeldaModelInfo(**model_json))
+            model_info.append(ZeldaModelInfo(**model_json))
+        
+        cls._model_infos = model_info
 
-    return models
+    @classmethod
+    def load_models(cls, path, **kwargs):
+        for model_info in cls._model_infos:
+            model = None
+
+            model_path = os.path.join(path, model_info.path + '.zip')
+            if os.path.exists(model_path):
+                model = PPO.load(model_path, **kwargs)
+            else:
+                model_directory = os.path.join(path, model_info.path)
+                if os.path.exists(model_directory):
+                    model_path = os.path.join(model_directory, 'best.zip')
+                    if os.path.exists(model_path):
+                        model = PPO.load(model_path, **kwargs)
+                    
+                    else:
+                        model_path = os.path.join(model_directory, 'model.zip')
+                        if os.path.exists(model_path):
+                            model = PPO.load(model_path, **kwargs)
+            
+            model_directory = os.path.join(path, model_info.path)
+            cls._models[model_info.name] = ZeldaModel(model_info, model)
+    
+    @classmethod
+    def get(cls, name : str) -> Optional['ZeldaModel']:
+        return cls._models.get(name, None)
+    
+    @classmethod
+    def get_all(cls) -> List['ZeldaModel']:
+        return list(cls._models.values())
