--- triforce_lib/scenarios.json
@@ -1,108 +1,0 @@
-{
-    "scenarios": [
-        {
-            "name": "dungeon1",
-            "description": "Dungeon 1 Full",
-            "critics": ["Dungeon1Critic"],
-            "end_conditions": ["Dungeon1EndCondition"],
-            "level" : 1,
-            "start": ["1_73"],
-            "data":
-            {
-                "bombs" : 0,
-                "keys" : 0,
-                "hearts_and_containers" : 34,
-                "partial_hearts" : 254
-            }
-        },
-        {
-            "name": "dungeon1beams",
-            "description": "Dungeon 1 Beams",
-            "critics": ["Dungeon1BeamCritic"],
-            "end_conditions": ["Dungeon1CombatEndCondition"],
-            "level" : 1,
-            "start": ["1_73"],
-            "data":
-            {
-                "bombs" : 0,
-                "keys" : 0,
-                "hearts_and_containers" : 34,
-                "partial_hearts" : 254
-            },
-            "fixed":
-            {
-                "hearts_and_containers" : 34,
-                "partial_hearts" : 254
-            }
-        },
-        {
-            "name": "dungeon1bombs",
-            "description": "Dungeon 1 Bombs",
-            "critics": ["Dungeon1BombCritic"],
-            "end_conditions": ["Dungeon1EndCondition"],
-            "level" : 1,
-            "start": ["1_73"],
-            "data":
-            {
-                "bombs" : 0,
-                "keys" : 0,
-                "hearts_and_containers" : 34,
-                "partial_hearts" : 254
-            },
-            "fixed":
-            {
-                "bombs" : 8
-            }
-        },
-        {
-            "name": "dungeon1nobeams",
-            "description": "Dungeon 1 combat trainer (no beams)",
-            "critics": ["Dungeon1Critic"],
-            "end_conditions": ["Dungeon1CombatEndCondition"],
-            "level" : 1,
-            "start": ["1_73"],
-            "data":
-            {
-                "bombs" : 0,
-                "keys" : 0,
-                "hearts_and_containers" : 33,
-                "partial_hearts" : 100
-            },
-            "fixed":
-            {
-                "hearts_and_containers" : 33,
-                "partial_hearts" : 100
-            }
-        },
-        {
-            "name": "dungeon1boss",
-            "description": "Dungeon 1 Boss",
-            "critics": ["Dungeon1BossCritic"],
-            "end_conditions": ["Dungeon1BossEndCondition"],
-            "level" : 1,
-            "start": ["1_35"],
-            "data":
-            {
-                "bombs" : 0,
-                "keys" : 0,
-                "hearts_and_containers" : 34,
-                "partial_hearts" : 254
-            }
-        },
-        {
-            "name": "dungeon1moveonly",
-            "description": "Dungeon 1 Move Only",
-            "critics": ["Dungeon1Critic"],
-            "end_conditions": ["Dungeon1CombatEndCondition"],
-            "level" : 1,
-            "start": ["1_73"],
-            "data":
-            {
-                "bombs" : 0,
-                "keys" : 8,
-                "hearts_and_containers" : 34,
-                "partial_hearts" : 254
-            }
-        }
-    ]
-}
